+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -125,21 +125,17 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
+                .Include(pt=>pt.Pets)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
-
-            return View(petType);
-        }
-
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
+            if  (petType.Pets.Count>0)
+            {
+                ModelState.AddModelError(string.Empty, "The petType can't be removed because existe pets of this pettype");
+                return RedirectToAction(nameof(Index));
+            }
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
